hahaha - need proj mtx
realized problem i was having with shadowing was same problem i had earlier.
because it is screen pass, the bgfx state u_view and u_proj are set for the screen quad,
not the scene rendered into gbuffer. need to store and pass separately as uniform.

but. just ran into this with view, can't believe i did it again with proj.

things kinda work now, but i need to fix shadow look.